got auto posting working
to test. go to facebook/index. press login to your facebook and enter your credentials, then click allow permissions and then click post to timeline. 

this is a big step. now we just need the facebook creditials stored in the database to be used to login and then we should be ready to publish to the virual girlfriend timeline